[FEAT] 잔액 조회 요청 API 기능 구현 및 테스트 코드 작성

+++ src/main/kotlin/kr/hhplus/be/server/domain/user/User.kt
@@ -15,7 +15,7 @@ class User(
 	var userUUID: String,
 
 	@Column(nullable = false, unique = true)
-	var balance: Long,
+	var balance: Int,
 ) : BaseEntity() {
 
 	fun updateUserUUID(uuid: String) {
